renamed graph to msgraph for api handler
