+++ apihandlers/msgraph/msgraph_api_handler_constants.go
@@ -1,5 +1,5 @@
-// graph_api_handler_constants.go
-package graph
+// msgraph_api_handler_constants.go
+package msgraph
 
 // Endpoint constants represent the URL suffixes used for graph API token interactions.
 const (
